restserver/requirements.txt
cachetools==3.1.1
Django==2.2.1
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
